ft_printf

+++ ft_printf.h
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   ft_printf.h                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: imatakis <@student.42abudhabi.ae>          +#+  +:+       +#+        */
+/*   By: imatakis <imatakis@student.42abudhabi.a    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/03/05 19:56:52 by imatakis          #+#    #+#             */
-/*   Updated: 2024/03/10 02:46:15 by imatakis         ###   ########.fr       */
+/*   Updated: 2024/03/11 19:25:29 by imatakis         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,6 +17,7 @@
 # include <stdarg.h>
 # include <stdio.h>
 # include <limits.h>
+# include <stdint.h>
 
 int	ft_printf(const char *s, ...);
 int print_char(int c);
@@ -24,6 +25,6 @@ int	print_unbr(unsigned int n);
 int	print_nbr(int n);
 int	print_str(const char *str);
 int	print_hex(unsigned int n, char c);
-int	print_ptr(unsigned int n);
+int	print_ptr(uintptr_t n);
 
 #endif
